+++ api_examples/rest_api/get_files_uuid_tags.rb
@@ -1,0 +1,4 @@
+#!/usr/bin/env ruby
+# frozen_string_literal: true
+
+require_relative '../support/run_rest_example'